refactor: change dir for login api files
