+++ src/feature/auth/login/api/use-logout.ts
@@ -1,7 +1,9 @@
 import { useContext } from 'react';
 import { useNavigate } from 'react-router';
+import { toast } from 'sonner';
 
 import AuthContext from '@/feature/auth/login/auth-provider';
+import { useCustomerStore } from '@/service/store/use-user-store';
 
 export function useLogout() {
   const { setIsAuthorized } = useContext(AuthContext);
@@ -10,7 +12,9 @@ export function useLogout() {
   const handleLogout = () => {
     localStorage.removeItem('ACCESS_TOKEN_KEY');
     setIsAuthorized(false);
+    useCustomerStore.getState().clearCustomer();
     void navigate('/login');
+    toast.success('Login successful!');
   };
 
   return handleLogout;
